removed deprecated symbol

+++ ags/test/audio/core-audio/core_audio_record.c
@@ -140,7 +140,7 @@ main(int argc, char **argv)
 			     sizeof(input_samplerate),
 			     &input_samplerate);
 
-  input_buffer_size_bytes = pcm_channels * buffer_size * sizeof(float);
+  input_buffer_size_bytes = (int) pcm_channels * buffer_size * sizeof(float);
 
   AudioObjectSetPropertyData(input_device,
 			     &input_buffer_size_property_address,
